+++ IOTV/src/lib/Protocols/iotvp_abstractreadwrite.h
@@ -7,7 +7,7 @@ class IOTVP_AbstractReadWrite : public IOTVP_AbstractBody
 public:
     virtual ~IOTVP_AbstractReadWrite() = default;
 
-    virtual uint32_t dataSize() const override;
+    virtual uint64_t dataSize() const override;
     const QByteArray &data() const;
 
     void setData(const QByteArray &newData);
